exercice1.3

+++ lab8_backend/routes/productRouter.js
@@ -4,37 +4,38 @@ const router = express.Router();
 // Le module multer sert à gérer les téléversements (upload) de fichiers
 const multer = require('multer');
 const storage = multer.memoryStorage();
-const upload = multer({ storage: storage });
+const upload = multer({storage: storage});
 
 const HttpError = require("../HttpError");
+const passport = require('passport');
 
 const productQueries = require("../queries/ProductQueries");
 
 // GET de la liste des produits
 // (Ne requiert pas d'authentification)
 router.get('/', (req, res, next) => {
-    productQueries.getAllProducts().then(products => {
-        res.json(products);
-    }).catch(err => {
-        return next(err);
-    });
+	productQueries.getAllProducts().then(products => {
+		res.json(products);
+	}).catch(err => {
+		return next(err);
+	});
 });
 
 
 // GET d'un produit individuel
 // (Ne requiert pas d'authentification)
 router.get('/:id', (req, res, next) => {
-    const id = req.params.id;
-    console.log("id:", id);
-    productQueries.getProductById(id).then(product => {
-        if (product) {
-            res.json(product);
-        } else {
-            return next(new HttpError(404, `Produit ${id} introuvable`));
-        }
-    }).catch(err => {
-        return next(err);
-    });
+	const id = req.params.id;
+	console.log("id:", id);
+	productQueries.getProductById(id).then(product => {
+		if (product) {
+			res.json(product);
+		} else {
+			return next(new HttpError(404, `Produit ${id} introuvable`));
+		}
+	}).catch(err => {
+		return next(err);
+	});
 });
 
 const onePixelTransparentPngImage = Buffer.from("iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAAAXNSR0IArs4c6QAAAA1JREFUGFdj+P///38ACfsD/QVDRcoAAAAASUVORK5CYII=", "base64");
@@ -42,23 +43,23 @@ const onePixelTransparentPngImage = Buffer.from("iVBORw0KGgoAAAANSUhEUgAAAAEAAAA
 // GET de l'image d'un produit
 // (Ne requiert pas d'authentification)
 router.get('/:id/image', (req, res, next) => {
-    const id = req.params.id;
-    console.log("id:", id);
-    productQueries.getProductImageContent(id).then(imageInfo => {
-        if (imageInfo && imageInfo.imageContent) {
-            if (imageInfo.imageContentType) {
-                res.header('Content-Type', imageInfo.imageContentType);
-            }
-            res.send(imageInfo.imageContent);
-        } else {
-            // Si le produit n'a pas d'image, on va retourner une image transparente de 1 pixel
-            // afin d'éviter d'avoir une image brisée dans le front-end
-            res.header('Content-Type', 'image/png');
-            res.send(onePixelTransparentPngImage);
-        }
-    }).catch(err => {
-        return next(err);
-    });
+	const id = req.params.id;
+	console.log("id:", id);
+	productQueries.getProductImageContent(id).then(imageInfo => {
+		if (imageInfo && imageInfo.imageContent) {
+			if (imageInfo.imageContentType) {
+				res.header('Content-Type', imageInfo.imageContentType);
+			}
+			res.send(imageInfo.imageContent);
+		} else {
+			// Si le produit n'a pas d'image, on va retourner une image transparente de 1 pixel
+			// afin d'éviter d'avoir une image brisée dans le front-end
+			res.header('Content-Type', 'image/png');
+			res.send(onePixelTransparentPngImage);
+		}
+	}).catch(err => {
+		return next(err);
+	});
 });
 
 
@@ -76,72 +77,82 @@ router.get('/:id/image', (req, res, next) => {
 //
 // Au besoin, référez-vous au module listeDifussionRouter.js dans l'exemple de code du cours 19.
 router.post('/',
-    (req, res, next) => {
-        const id = req.body.id;
-        if (!id || id === '') {
-            // Le return fait en sorte qu'on n'exécutera pas le reste de la fonction
-            // après l'appel à next(...).
-            return next(new HttpError(400, 'Le champ id est requis'));
-        }
+	passport.authenticate('basic', {session: false}),
+	(req, res, next) => {
+		if (!req.user.isAdmin) {
+			return next(new HttpError(403, 'Vous n\'avez pas les permissions'));
+		}
 
-        productQueries.getProductById(id).then(product => {
-            if (product) {
-                throw new HttpError(400, `Un produit avec l'id ${id} existe déjà`);
-            }
+		const id = req.body.id;
+		if (!id || id === '') {
+			// Le return fait en sorte qu'on n'exécutera pas le reste de la fonction
+			// après l'appel à next(...).
+			return next(new HttpError(400, 'Le champ id est requis'));
+		}
 
-            const newProduct = {
-                id: "" + id,
-                name: "" + req.body.name,
-                price: + req.body.price,
-                desc: "" + req.body.desc,
-                image: "" + req.body.image,
-                longDesc: "" + req.body.longDesc
-            };
+		productQueries.getProductById(id).then(product => {
+			if (product) {
+				throw new HttpError(400, `Un produit avec l'id ${id} existe déjà`);
+			}
 
-            return productQueries.insertProduct(newProduct);
-        }).then(result => {
-            res.json(result);
-        }).catch(err => {
-            next(err);
-        });
+			const newProduct = {
+				id: "" + id,
+				name: "" + req.body.name,
+				price: +req.body.price,
+				desc: "" + req.body.desc,
+				image: "" + req.body.image,
+				longDesc: "" + req.body.longDesc
+			};
 
-    });
+			return productQueries.insertProduct(newProduct);
+		}).then(result => {
+			res.json(result);
+		}).catch(err => {
+			next(err);
+		});
+
+	});
 
 // PUT pour la modification d'un produit
 // ** Exercice 1.3 **
 // Approche similaire que pour le POST ci-haut. La modification d'un produit
 // doit être refusée pour les comptes non-administrateurs (avec un statut HTTP 403).
 router.put('/:id',
-    (req, res, next) => {
-        const id = req.params.id;
-        if (!id || id === '') {
-            return next(new HttpError(400, 'Le paramètre id est requis'));
-        }
+	passport.authenticate('basic', {session: false}),
+	(req, res, next) => {
+		if (!req.user.isAdmin) {
+			return next(new HttpError(403, 'Vous n\'avez pas les permissions'));
+		}
 
-        if (id !== req.body.id) {
-            return next(new HttpError(400, `Le paramètre spécifie l'id ${id} alors que le produit fourni a l'id ${req.body.id}`));
-        }
+		const id = req.params.id;
+		if (!id || id === '') {
+			return next(new HttpError(400, 'Le paramètre id est requis'));
+		}
 
-        const newProduct = {
-            id: "" + id,
-            name: "" + req.body.name,
-            price: + req.body.price,
-            desc: "" + req.body.desc,
-            image: "" + req.body.image,
-            longDesc: "" + req.body.longDesc
-        };
+		if (id !== req.body.id) {
+			return next(new HttpError(400, `Le paramètre spécifie l'id ${id} alors que le produit fourni a l'id ${req.body.id}`));
+		}
 
-        productQueries.updateProduct(newProduct).then(result => {
-            if (!result) {
-                return next(new HttpError(404, `Produit ${id} introuvable`));
-            }
+		const newProduct = {
+			id: "" + id,
+			name: "" + req.body.name,
+			price: +req.body.price,
+			desc: "" + req.body.desc,
+			image: "" + req.body.image,
+			longDesc: "" + req.body.longDesc
+		};
 
-            res.json(result);
-        }).catch(err => {
-            return next(err);
-        });
+		productQueries.updateProduct(newProduct).then(result => {
+			if (!result) {
+				return next(new HttpError(404, `Produit ${id} introuvable`));
+			}
 
-    });
+			res.json(result);
+		}).catch(err => {
+			return next(err);
+		});
+
+	});
 
 
 // DELETE pour le retrait d'un produit (pas utilisé par le front-end actuel)
@@ -149,54 +160,59 @@ router.put('/:id',
 // Approche similaire que pour le POST ci-haut. Le retrait d'un produit
 // doit être refusée pour les comptes non-administrateurs (avec un statut HTTP 403).
 router.delete('/:id',
-    (req, res, next) => {
-        const id = req.params.id;
-        if (!id || id === '') {
-            return next(new HttpError(400, 'Le paramètre id est requis'));
-        }
+	passport.authenticate('basic', {session: false}),
+	(req, res, next) => {
+		if (!req.user.isAdmin) {
+			return next(new HttpError(403, 'Vous n\'avez pas les permissions'));
+		}
 
-        productQueries.deleteProduct(id).then(result => {
-            if (!result) {
-                return next(new HttpError(404, `Produit ${id} introuvable`));
-            }
+		const id = req.params.id;
+		if (!id || id === '') {
+			return next(new HttpError(400, 'Le paramètre id est requis'));
+		}
 
-            res.json(result);
-        }).catch(err => {
-            return next(err);
-        });
-    });
+		productQueries.deleteProduct(id).then(result => {
+			if (!result) {
+				return next(new HttpError(404, `Produit ${id} introuvable`));
+			}
+
+			res.json(result);
+		}).catch(err => {
+			return next(err);
+		});
+	});
 
 // POST de l'image d'un produit
 // ** Exercice 1.3 **
 // Approche similaire que pour le POST ci-haut. Le changement d'image d'un produit
 // doit être refusé pour les comptes non-administrateurs (avec un statut HTTP 403).
 router.post('/:id/image',
-    // Fonction middleware de multer pour gérer l'upload d'un fichier dans ce endpoint.
-    // Cet appel de middleware doit venir après celui de l'authentification.
-    upload.single('product-image'), // doit correspondre à l'id du champ dans le formulaire html
-    (req, res, next) => {
-        const id = req.params.id;
-        if (!id || id === '') {
-            // Le return fait en sorte qu'on n'exécutera pas le reste de la fonction
-            // après l'appel à next(...).
-            return next(new HttpError(400, 'Le champ id est requis'));
-        }
+	// Fonction middleware de multer pour gérer l'upload d'un fichier dans ce endpoint.
+	// Cet appel de middleware doit venir après celui de l'authentification.
+	upload.single('product-image'), // doit correspondre à l'id du champ dans le formulaire html
+	(req, res, next) => {
+		const id = req.params.id;
+		if (!id || id === '') {
+			// Le return fait en sorte qu'on n'exécutera pas le reste de la fonction
+			// après l'appel à next(...).
+			return next(new HttpError(400, 'Le champ id est requis'));
+		}
 
-        productQueries.getProductById(id).then(product => {
-            if (!product) {
-                throw new HttpError(404, `Produit id ${id} introuvable`);
-            }
+		productQueries.getProductById(id).then(product => {
+			if (!product) {
+				throw new HttpError(404, `Produit id ${id} introuvable`);
+			}
 
-            // Le middleware upload.single(...) rendra accessible le contenu binaire du fichier
-            // téléversé dans req.file.buffer et le type de fichier (p.ex. "image/jpeg")
-            // dans req.file.mimetype:
-            return productQueries.updateProductImage(id, req.file.buffer, req.file.mimetype);
-        }).then(imageInfo => {
-            res.send("");
-        }).catch(err => {
-            next(err);
-        });
+			// Le middleware upload.single(...) rendra accessible le contenu binaire du fichier
+			// téléversé dans req.file.buffer et le type de fichier (p.ex. "image/jpeg")
+			// dans req.file.mimetype:
+			return productQueries.updateProductImage(id, req.file.buffer, req.file.mimetype);
+		}).then(imageInfo => {
+			res.send("");
+		}).catch(err => {
+			next(err);
+		});
 
-    });
+	});
 
 module.exports = router;
